// src/models/RegistroInterface.ts
// src/models/user.ts
export interface RegistroInterface {
    id: number;
    fecha: string;
    hora: string;
    ip: string;
    conectada: boolean;
    numSerie: string;
    modelo: string;
    contador: number;
    negro: number;
    color:boolean;
    cyan: number;
    amarillo: number;
    rojo: number;
  }